requirements.txt
backpack==0.1
bcrypt==3.2.0
blinker==1.4
boto3==1.20.13
botocore==1.23.13
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
cleo==0.6.8
click==8.0.3
Faker==0.8.18
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT==0.3.2
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-vision==2.6.2
googleapis-common-protos==1.53.0
grpcio==1.42.0
grpcio-status==1.42.0
idna==3.3
importlib-metadata==4.8.2
inflection==0.3.1
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
orator==0.9.9
pastel==0.1.1
pdf2image==1.16.0
pendulum==1.5.1
proto-plus==1.19.8
protobuf==3.19.1
pyaml==16.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.10.0
PyJWT==1.4.2
pylev==1.4.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv-yaml==0.17.1
pytz==2021.3
pytzdata==2020.1
PyYAML==5.4.1
pyzbar==0.1.9
requests==2.26.0
rsa==4.8
s3transfer==0.5.0
simplejson==3.17.6
six==1.16.0
text-unidecode==1.2
textract-trp==0.1.3
typing_extensions==4.0.0
tzlocal==1.5.1
urllib3==1.26.7
uWSGI==2.0.20
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
pytesseract